- Biometrics.Repair method created for performing repair steps. - Player.Minerals and Player.VespeneGas changed from int to RCNumber to support fractional amounts during repair. - SCVRepairExecution implemented. - Backlog item RC10BI_SCV_BASICS done. - New solution created for prototype projects at /proto, some old prototypes removed. - Initial version of a prototype created for the new pathfinding.

+++ src/RC.App.BizLogic/Views/Core/PlayerView.cs
@@ -28,7 +28,7 @@ namespace RC.App.BizLogic.Views.Core
             get
             {
                 if (this.Scenario == null) { throw new InvalidOperationException("No active scenario!"); }
-                return this.Scenario.GetPlayer((int)this.selectionManager.LocalPlayer).Minerals;
+                return (int)this.Scenario.GetPlayer((int)this.selectionManager.LocalPlayer).Minerals;
             }
         }
 
@@ -38,7 +38,7 @@ namespace RC.App.BizLogic.Views.Core
             get
             {
                 if (this.Scenario == null) { throw new InvalidOperationException("No active scenario!"); }
-                return this.Scenario.GetPlayer((int)this.selectionManager.LocalPlayer).VespeneGas;
+                return (int)this.Scenario.GetPlayer((int)this.selectionManager.LocalPlayer).VespeneGas;
             }
         }
